requirements.txt
Flask>=2.0,<3.0
selenium>=4.0
webdriver-manager>=4.0
python-dotenv>=0.21
waitress>=2.2
requests>=2.31
